+++ constants/__init__.py
@@ -1,0 +1,13 @@
+# As of today the object build limit for bases is 3k
+MAX_BASE_OBJS = 3000
+
+# Object IDs of various base parts
+BASE_FLAG_ID = "^BASE_FLAG"
+STONE_DOME_ROOF = "^S_ROOF5"
+WOOD_FLOOR_TILE = "^T_FLOOR"
+STONE_FLOOR_TILE = "^S_FLOOR"
+PAVING = "^BUILDPAVING_BIG"
+
+
+# DEFAULT_OBJECT is the "standard" or "background" used when mapping
+DEFAULT_OBJECT = STONE_FLOOR_TILE
